Replaces the name gemm by matrix_multiply
- gemm is the interface from BLAS. Since the interface in T-LAPACK is different, it makes sense to rename it.
- Also, by renaming it we could make different design choices.

`matrix_multiply` has two versions: one with and one without beta.
1. The first version computes `C := alpha*A*B + beta*C`;
2. The second computes `C := alpha*A*B`.
They are semantically different: in the first, C is [in,out]; in the second, C is [out] only.
Legacy `gemm` covers both functionalities depending on the value of beta (if beta is zero, use second expression).
The problem with the strategy adopted in `gemm` is: if beta evaluates to zero in some part of an algorithm and there are Infs or NaNs in C, those will not propagate.

The proposal is: if one wants to obtain the same behavior of gemm and
- `alpha=0` and `beta=0`: just use `laset` to set the matrix C to 0.
- `alpha=0` and `beta` is not zero: just use `lascl` to scale the matrix C.
- `alpha` is not zero and `beta=0`: use the version of `matrix_multiply` that considers C as output only.
- otherwise, use `matrix_multiply` that considers C as input and output.

This commit also introduces the StrongZero internal data type. This is to avoid code duplication.