Example with StarPU matrices is complete
New namespace tlapack::starpu implements:
- Matrix.
  - It has a specific way to create grids. Still not completely
functional. Currently only works if the sizes of the matrix are multiple
of the partition sizes.
  - A virtual way to create submatrices. It means: we don't change the
handle when extracting submatrices.
  - Access operators operator() and operator[]. The second only works
for 'vectors'
  - Vectors: implemented under the same matrix structure. It checks at
runtime if a type is a matrix or vector depending on the operation.
- Matrix<T>::data is the non-const data type of each entry. It is
different from T. Matrix<T>::data is associated to StarPU handles.
  - Operations: { Assign, Add, Subtract, Multiply, Divide }. Operations
between (1) two Matrix<T>::data, and (2) one Matrix<T>::data and T. All
operations are done using StarPU tasks.
  - abs and sqrt are as well implemented.
  - Implicit conversion from Matrix<T>::data to T is also implemented.